moved enums to enum subdir
