+++ weatherterm/core/__init__.py
@@ -1,8 +1,6 @@
-from .base_enum import BaseEnum
 from .forecast import Forecast
-from .forecast_type import ForecastType
+from .mapper import Mapper
+# parser_loader isn't a class
 from .request import Request
-from .unit import Unit
 from .set_unit_action import SetUnitAction
 from .unit_converter import UnitConverter
-from .mapper import Mapper
