+++ weatherterm/core/enum/base_enum.py
