+++ weatherterm/core/enum/forecast_type.py
