+++ weatherterm/core/set_unit_action.py
@@ -1,5 +1,5 @@
 from argparse import Action
-from weatherterm.core import Unit
+from .enum.unit import Unit
 
 class SetUnitAction(Action):
     def __call__(self, parser, namespace, values, option_string = None):
