+++ weatherterm/core/unit_converter.py
@@ -1,4 +1,4 @@
-from .unit import Unit
+from .enum.unit import Unit 
 
 class UnitConverter:
     # Converts temperature units between Celsius & Fahrenheit
